Removed unnecessary code, code organized, cursor changed, manifest file updated
