+++ public/index.html
@@ -6,14 +6,14 @@
   <link rel="icon" href="%PUBLIC_URL%/favicon.png" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <meta name="theme-color" content="#000000" />
-  <meta name="description" content="Ajaysinh Rathod - Software Engineer | Mobile and Web Development | UI Design" /> <!-- Updated description with additional keywords -->
-  <meta name="keywords" content="Ajaysinh Rathod, Software Engineer, Mobile Development, Web Development, UI Design, Portfolio" /> <!-- Added keywords -->
-  <meta name="author" content="Ajaysinh Rathod" /> <!-- Added author's name -->
+  <meta name="description" content="Ajaysinh Rathod - Software Engineer | Mobile and Web Development | UI Design" /> 
+  <meta name="keywords" content="Ajaysinh Rathod, Software Engineer, Mobile Development, Web Development, UI Design, Portfolio" />
+  <meta name="author" content="Ajaysinh Rathod" />
   <link rel="preconnect" href="https://fonts.googleapis.com" />
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
   <link href="https://fonts.googleapis.com/css2?family=IBM+Plex+Mono:wght@400;500;600&family=Michroma&family=Montserrat:wght@600;700&display=swap" rel="stylesheet" />
   <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
-  <title>Ajaysinh Rathod - Software Engineer</title> <!-- Updated title with profession -->
+  <title>Ajaysinh Rathod - Software Engineer</title>
 </head>
 
 <body>
