+++ src/App.css
@@ -1,4 +1,5 @@
 .App {
   text-align: center;
   position: relative;
+
 }
