+++ src/components/about/About.js
@@ -1,4 +1,4 @@
-import SectionHeading from '../custom_components/SectionHeading';
+import SectionHeading from '../custom_components/section_heading/SectionHeading';
 import './About.css';
 
 function About() {
