+++ src/components/contact/Contact.css
@@ -1,17 +1,24 @@
 #sayHelloButton {
   text-decoration: none;
-  margin: 0vw 1vw;
-  padding: 2vh 3vw;
+  margin: 1vw 0vw;
+  padding: 1vw 3vw;
   border: 1px solid #000;
   box-shadow: 6px 6px 0px 0px var(--box-shadow-color);
   color: #000;
-  font-size: 20px;
+  display: flex;
+  gap: 5px;
+  align-items: center;
+  font-size: 22px;
+  justify-content: center;
 }
-
+#leetcode{
+  height: 45px;
+}
 #contacttitle {
   text-transform: lowercase;
-  font-size: 32px;
+  font-size: 1.8vw;
   margin: 0vw 4vw;
+  width: 80%;
 }
 #contactsection {
   display: flex;
@@ -24,7 +31,7 @@
 
 #socialmedialinks {
   width: 100%;
-  margin-top: 10vh;
+  margin-top: 2vw;
   display: flex;
   flex-direction: row;
   align-items: center;
@@ -33,23 +40,39 @@
   background-color: #000;
   padding: 4vw 0vw;
 }
+#socialmedialinks > a {
 
-@media (max-width: 800px) {
+}
+@media (max-width: 1200px) {
   #contacttitle {
-    font-size: 24px;
+    font-size: 3vw;
   }
   #sayHelloButton {
-    padding: 2vh 4vw;
+    width: 20%;
+    padding: 2vw 1vw;
+  }
+}
+@media (max-width: 800px) {
+  #contacttitle {
+    font-size: 4vw;
+  }
+  #sayHelloButton {
+    width: 30%;
     font-size: 20px;
+    height: 40px;
+    padding: 2vw 1vw;
   }
 }
 @media (max-width: 600px) {
   #contacttitle {
     font-size: 20px;
+
   }
   #sayHelloButton {
-    padding: 2vh 5vw;
+    width: 60%;
     font-size: 18px;
+    height: 40px;
+    padding: 2vw 1vw;
   }
   #socialmedialinks {
     margin-top: 4vh;
@@ -58,6 +81,6 @@
   }
   #socialmedialinks > a > img {
     width: 28px;
-    height: 28px; 
+    height: 28px;
   }
 }
