--- src/components/custom_components/Loading.css
@@ -1,55 +1,0 @@
-.loading-screen {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: #fff;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    z-index: 1000;
-    transition: opacity 0.5s ease-in-out;
-    opacity: 1;
-  }
-  
-  .loading-icon {
-    width: 60px;
-    height: 60px;
-    border: 2px solid transparent;
-    border-right: 2px solid #000000;
-    border-radius: 0%;
-    animation: borderAnimation 2s linear infinite;
-    text-align: center;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-  }
-  
-  @keyframes borderAnimation {
-    0%, 100% {
-      border-right: 2px solid #000000;
-      border-bottom: 2px solid transparent;
-      border-left: 2px solid transparent;
-      border-top: 2px solid transparent;
-    }
-    25% {
-      border-right: 2px solid transparent;
-      border-bottom: 2px solid #000000;
-      border-left: 2px solid transparent;
-      border-top: 2px solid transparent;
-    }
-    50% {
-      border-right: 2px solid transparent;
-      border-bottom: 2px solid transparent;
-      border-left: 2px solid #000000;
-      border-top: 2px solid transparent;
-    }
-    75% {
-      border-right: 2px solid transparent;
-      border-bottom: 2px solid transparent;
-      border-left: 2px solid transparent;
-      border-top: 2px solid #000000;
-    }
-  }
-
