--- src/components/custom_components/Loading.js
@@ -1,13 +1,0 @@
-import React from "react";
-import "./Loading.css";
-// Assuming the favicon.png is in the public directory
-
-function LoadingScreen() {
-    return (
-        <div className="loading-screen">
-            <div className="loading-icon" >0100 <br /> 0001</div>
-        </div>
-    );
-}
-
-export default LoadingScreen;
