+++ src/components/hero/Hero.css
@@ -4,9 +4,12 @@
   font-family: "Michroma";
   font-size: 4vw;
 }
-
+#firstnamesection {
+  margin-top: 12vw;
+}
 #hero {
-  margin: 26vh 0vw;
+  margin: 20vh 0vw;
+  margin-bottom: 10vw;
 }
 
 #designation {
@@ -22,6 +25,7 @@
   margin-left: -12.76vw;
   overflow: hidden;
   display: flex;
+  mix-blend-mode: difference;
   justify-content: start;
 }
 
@@ -59,7 +63,7 @@
   color: #fff;
 }
 #actionbuttons {
-  margin: 15vh 1vw;
+  margin: 8vw 1vw;
   display: flex;
   justify-content: center;
   align-items: end;
@@ -122,7 +126,7 @@
 
   #actionbuttons > a {
     font-size: 14px;
-    margin: 0vw 2vw;
+    margin: 12vw 2vw;
   }
   #viewresume {
     padding: 14px 26px;
