+++ src/components/navbar/NavBar.css
@@ -18,7 +18,7 @@
 .navbar > a {
     text-decoration: none;
     color: #000;
-    cursor: pointer;
+
 }
 
 #menuicon {
