+++ src/components/projects/Projects.js
@@ -1,4 +1,4 @@
-import SectionHeading from "../custom_components/SectionHeading";
+import SectionHeading from "../custom_components/section_heading/SectionHeading";
 import NineTeenFiveImage from "../../assets/nineteenfive.webp";
 import SanskarPgImage from "../../assets/sanskar-pg-app.webp";
 import ListlyImage from "../../assets/listly.webp";
